fixed: ":project_so not found error" when :platform_configuration is not set
This is actually has  to be done in 7ebec9f (improve: paclient: respect
configuration, 2015-03-15).
When :platform_configuration is not set, then deployment files list is
actually empty, that leads to the confusing "project_so not found"
error.
To avoid such a confusing behaviour, let's assume 'Debug'
:platform_configuration implicitly set and also show warnings when
deployment files list is empty. Besides, show an advice to check
:platform_configuration in then "project_so not found" error.