fixed: .map-file does not contain details if debuginfo is not set
If debuginfo is not set (and even if we set debug=true and/or
map=detail) we can get .map file without actual detailed debug info.
Why? Because setting :debug to `true` passes
'-$D+ -$L+ -$YD -$C+ -$Q+ -$R+ -$O- -GD' to dcc32.exe
and :debuginfo THEN sets/unsets '-$D+ -$L+ -$YD'/'-$D- -$L- -$Y-'

So, we MUST always set debuginfo even if we set :debug. That is annoying
thing. Setting :debug would be enough to get a debug version with a
detailed .map-file.

So, :debuginfo has to have THREE states: nil/not set/set but
not: "not set"/set.